Add AlienDpkg Provider.
This commit adds an AlienDpkg provider for the package resource. The
main purpose of this provider is to automatically substitute an rpm
package source path with a debian source path.  Also if the debian
package has not been generated, the provider automatically attempts to
convert the rpm package to a debian package.